move createemptyconstraint out
